notifier_api/requirements.txt
Django
djangorestframework
PyPI
six
psycopg2
appdirs
packaging
pyparsing
pytz
psycopg2-binary
